+++ ElvUI_EltreumUI/Modules/Misc/QuestCombat.lua
@@ -20,7 +20,8 @@ function ElvUI_EltreumUI:QuestEncounter()
 			if E.Retail then
 				--ObjectiveTracker_Collapse()
 				--ObjectiveTrackerFrame:Hide()
-				ObjectiveTrackerFrame:SetAlpha(0)
+				ObjectiveTrackerFrame:SetParent(E.HiddenFrame)
+				--ObjectiveTrackerFrame:SetAlpha(0)
 			elseif E.Classic then
 				if IsAddOnLoaded("Questie") then
 					if _G["Questie_BaseFrame"] then
@@ -52,7 +53,8 @@ function ElvUI_EltreumUI:QuestEncounterEnd()
 		local _, instanceType = IsInInstance()
 		if instanceType == "raid" or instanceType == "party" or instanceType == "scenario" then --and event == "PLAYER_REGEN_DISABLED"
 			if E.Retail then
-				ObjectiveTrackerFrame:SetAlpha(1)
+				ObjectiveTrackerFrame:SetParent(UIParent)
+				--ObjectiveTrackerFrame:SetAlpha(1)
 			elseif E.Classic then
 				if IsAddOnLoaded("Questie") then
 					if _G["Questie_BaseFrame"] then
@@ -85,7 +87,8 @@ function ElvUI_EltreumUI:ArenaQuest()
 		if instanceType == "arena" or instanceType == "pvp" then
 			if E.Retail then
 				--ObjectiveTrackerFrame:Hide()
-				ObjectiveTrackerFrame:SetAlpha(0)
+				ObjectiveTrackerFrame:SetParent(E.HiddenFrame)
+				--ObjectiveTrackerFrame:SetAlpha(0)
 			elseif E.Classic then
 				if IsAddOnLoaded("Questie") then
 					if _G["Questie_BaseFrame"] then
@@ -109,7 +112,8 @@ function ElvUI_EltreumUI:ArenaQuest()
 			end
 		elseif instanceType == "none" then
 			if E.Retail then
-				ObjectiveTrackerFrame:SetAlpha(1)
+				ObjectiveTrackerFrame:SetParent(UIParent)
+				--ObjectiveTrackerFrame:SetAlpha(1)
 			elseif E.Classic then
 				if IsAddOnLoaded("Questie") then
 					if _G["Questie_BaseFrame"] then
@@ -143,7 +147,8 @@ function ElvUI_EltreumUI:QuestCombat()
 				_G.ObjectiveTrackerFrame:ToggleCollapsed() --seems like this causes taints
 			end]]
 			--ObjectiveTrackerFrame:SetCollapsed(true)
-			ObjectiveTrackerFrame:SetAlpha(0)
+			--ObjectiveTrackerFrame:SetAlpha(0)
+			ObjectiveTrackerFrame:SetParent(E.HiddenFrame)
 			--ObjectiveTracker_Collapse()
 		elseif E.Classic then
 			if IsAddOnLoaded("Questie") then
@@ -177,7 +182,8 @@ function ElvUI_EltreumUI:QuestCombatEnd()
 				_G.ObjectiveTrackerFrame:ToggleCollapsed() --seems like this causes taints
 			end]]
 			--ObjectiveTrackerFrame:SetCollapsed(false)
-			ObjectiveTrackerFrame:SetAlpha(1)
+			--ObjectiveTrackerFrame:SetAlpha(1)
+			ObjectiveTrackerFrame:SetParent(UIParent)
 			--ObjectiveTracker_Expand()
 		elseif E.Classic then
 			if IsAddOnLoaded("Questie") then