+++ ElvUI_EltreumUI/Modules/Skins/Addons/Hekili.lua
@@ -9,6 +9,7 @@ function ElvUI_EltreumUI:EltruismHekili()
 			for i = 1, 10 do
 				if _G["Hekili_Primary_B"..i] then
 					S:HandleButton(_G["Hekili_Primary_B"..i])
+					_G["Hekili_Primary_B"..i].Texture:SetAlpha(_G["Hekili_Primary_B"..i]:GetAlpha())
 					if E.db.ElvUI_EltreumUI.skins.shadow.enable and not _G["Hekili_Primary_B"..i].shadow then
 						_G["Hekili_Primary_B"..i]:CreateShadow(E.db.ElvUI_EltreumUI.skins.shadow.length)
 						ElvUI_EltreumUI:ShadowColor(_G["Hekili_Primary_B"..i].shadow)
@@ -22,6 +23,7 @@ function ElvUI_EltreumUI:EltruismHekili()
 				end
 				if _G["Hekili_Defensives_B"..i] then
 					S:HandleButton(_G["Hekili_Defensives_B"..i])
+					_G["Hekili_Defensives_B"..i].Texture:SetAlpha(_G["Hekili_Defensives_B"..i]:GetAlpha())
 					if E.db.ElvUI_EltreumUI.skins.shadow.enable and not _G["Hekili_Defensives_B"..i].shadow then
 						_G["Hekili_Defensives_B"..i]:CreateShadow(E.db.ElvUI_EltreumUI.skins.shadow.length)
 						ElvUI_EltreumUI:ShadowColor(_G["Hekili_Defensives_B"..i].shadow)
@@ -35,6 +37,7 @@ function ElvUI_EltreumUI:EltruismHekili()
 				end
 				if _G["Hekili_Cooldowns_B"..i] then
 					S:HandleButton(_G["Hekili_Cooldowns_B"..i])
+					_G["Hekili_Cooldowns_B"..i].Texture:SetAlpha(_G["Hekili_Cooldowns_B"..i]:GetAlpha())
 					if E.db.ElvUI_EltreumUI.skins.shadow.enable and not _G["Hekili_Cooldowns_B"..i].shadow then
 						_G["Hekili_Cooldowns_B"..i]:CreateShadow(E.db.ElvUI_EltreumUI.skins.shadow.length)
 						ElvUI_EltreumUI:ShadowColor(_G["Hekili_Cooldowns_B"..i].shadow)
@@ -48,6 +51,7 @@ function ElvUI_EltreumUI:EltruismHekili()
 				end
 				if _G["Hekili_AOE_B"..i] then
 					S:HandleButton(_G["Hekili_AOE_B"..i])
+					_G["Hekili_AOE_B"..i].Texture:SetAlpha(_G["Hekili_AOE_B"..i]:GetAlpha())
 					if E.db.ElvUI_EltreumUI.skins.shadow.enable and not _G["Hekili_AOE_B"..i].shadow then
 						_G["Hekili_AOE_B"..i]:CreateShadow(E.db.ElvUI_EltreumUI.skins.shadow.length)
 						ElvUI_EltreumUI:ShadowColor(_G["Hekili_AOE_B"..i].shadow)
@@ -61,6 +65,7 @@ function ElvUI_EltreumUI:EltruismHekili()
 				end
 				if _G["Hekili_Interrupts_B"..i] then
 					S:HandleButton(_G["Hekili_Interrupts_B"..i])
+					_G["Hekili_Interrupts_B"..i].Texture:SetAlpha(_G["Hekili_Interrupts_B"..i]:GetAlpha())
 					if E.db.ElvUI_EltreumUI.skins.shadow.enable and not _G["Hekili_Interrupts_B"..i].shadow then
 						_G["Hekili_Interrupts_B"..i]:CreateShadow(E.db.ElvUI_EltreumUI.skins.shadow.length)
 						ElvUI_EltreumUI:ShadowColor(_G["Hekili_Interrupts_B"..i].shadow)