+++ Changelog.md
@@ -1,6 +1,17 @@
 ### 🌌 Next Version
 The future is uncertain
 ___
+### 4.1.9.1
+Added a menu in Eltruism > Installer to import and export Gradient settings  
+Updated Experimental Thin Mode layout to add a new layout instead of Overwriting current one
+Updated Border Frame Levels to allow higher levels  
+Updated how the Quest Tracker hides  
+Fixed Pet Autocast error in Classic/Cataclysm due to 11.0 changes  
+Fixed icon for professions in Classic and Cataclysm  
+Fixed a possible error with the Quest Tracker skin  
+Fixed Hekili's skin hiding the Textures  
+Fixed Event Toast lines  
+Fixed Game Menu Button  
 ### 4.1.9
 Added support for The War Within  
 Added several more statusbar textures from Blizzard's default textures